+++ src/main/java/racingcar/domain/App.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.view.InputView;
 
 public class App {
     public static void main(String[] args) {
