+++ src/main/java/racingcar/domain/Game.java
@@ -1,4 +1,7 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.view.LaneViewModel;
+import racingcar.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
